[Http] Move HeaderValue constant class to Http/Message from Auth.
